Add a new record with auto increment id

+++ SpringDataElasticSearchDemo/src/main/java/com/demo/repository/BookRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 
 import com.demo.entity.Book;
 
-public interface BookRepository extends ElasticsearchRepository<Book, Integer> {
+public interface BookRepository extends ElasticsearchRepository<Book, Double> {
 
 }
